멱등성 수정, 부하테스트 수정, 장바구니 매장별로 들어가게 수정

+++ src/main/java/com/whale/order/domain/cart/dto/CartAddRequest.java
@@ -15,6 +15,6 @@ public record CartAddRequest(
             Long menuOptionId,
             String optionGroup,
             String optionName,
-            Integer additionalPrice
+            Long additionalPrice
     ) {}
 }
