+++ src/main/java/com/whale/order/domain/cart/dto/CartItem.java
@@ -20,11 +20,11 @@ public class CartItem {
     private Long menuId;
     private String menuName;
     private String imageUrl;
-    private Integer basePrice;
+    private Long basePrice;
     private Integer quantity;
     private List<SelectedOption> selectedOptions;
-    private Integer unitPrice;   // basePrice + 옵션 추가금 합산
-    private Integer totalPrice;  // unitPrice * quantity
+    private Long unitPrice;   // basePrice + 옵션 추가금 합산
+    private Long totalPrice;  // unitPrice * quantity
 
     @Getter
     @NoArgsConstructor
@@ -34,6 +34,6 @@ public class CartItem {
         private Long menuOptionId;
         private String optionGroup;
         private String optionName;
-        private Integer additionalPrice;
+        private Long additionalPrice;
     }
 }
