+++ src/main/java/com/whale/order/domain/cart/dto/CartResponse.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record CartResponse(
         List<CartItem> items,
-        int totalPrice,
+        long totalPrice,
         int totalCount
 ) {}
