+++ src/main/java/com/whale/order/domain/menu/dto/CustomerMenuResponse.java
@@ -9,7 +9,7 @@ public record CustomerMenuResponse(
         Long menuId,
         String name,
         String description,
-        Integer basePrice,
+        Long basePrice,
         String category,
         String imageUrl,
         LocalDate saleStartDate,
