+++ src/main/java/com/whale/order/domain/menu/dto/MenuDetailResponse.java
@@ -10,7 +10,7 @@ public record MenuDetailResponse(
         Long menuId,
         String name,
         String description,
-        Integer basePrice,
+        Long basePrice,
         String category,
         String imageUrl,
         LocalDate saleStartDate,
