+++ src/main/java/com/whale/order/domain/menu/dto/MenuResponse.java
@@ -8,7 +8,7 @@ public record MenuResponse(
         Long menuId,
         String name,
         String description,
-        Integer basePrice,
+        Long basePrice,
         String category,
         String imageUrl,
         LocalDate saleStartDate,
