+++ src/main/java/com/whale/order/domain/menu/dto/MenuUpdateRequest.java
@@ -25,7 +25,7 @@ public class MenuUpdateRequest {
 
     @NotNull(message = "기본 가격은 필수입니다.")
     @Min(value = 0, message = "가격은 0원 이상이어야 합니다.")
-    private Integer basePrice;
+    private Long basePrice;
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
     private LocalDate saleStartDate;
